third_party/WebKit/LayoutTests/svg/filters/feBlend-invalid-mode-expected.txt

CONSOLE MESSAGE: line 14: Uncaught SVG_INVALID_VALUE_ERR: An invalid value was passed to an operation or assigned to an attribute.
PASS: Invalid blend modes do not trigger a crash. 
